pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lagrangian-uv-stage2"
version = "0.1.0"
description = "Stage-2 stochastic-advection state-space model for 140m UV wind components."
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "PyYAML>=6.0",
    "torch>=2.2",
    "h5py>=3.10",
    "scipy>=1.13",
    "matplotlib>=3.8",
    "imageio>=2.34",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
